added 'typings' file to hold all custom TS types/interfaces
